% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildHttrDB.R
\name{buildHttrDB}
\alias{buildHttrDB}
\title{BUild the mysql database for the httr data}
\usage{
buildHttrDB(
  do.clean = F,
  do.init = F,
  do.load.datasets = F,
  do.load.pods = F,
  sigcatalog = "signatureDB_master_catalog 2022-05-16",
  sigset = "screen_large",
  method = "gsea",
  hccut = 0.9,
  tccut = 1,
  user = "rjudson",
  password
)
}
\description{
BUild the mysql database for the httr data
}
