% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfasQC.tp.db.R
\name{pfasQC.tp.db}
\alias{pfasQC.tp.db}
\title{BUild the mysql database for the httr data}
\usage{
pfasQC.tp.db(
  do.clean = F,
  do.init = F,
  do.load.datasets = F,
  do.load.pods = F,
  sigcatalog = "signatureDB_master_catalog 2021-09-29",
  sigset = "screen_large",
  method = "gsea",
  hccut = 0.9,
  tccut = 1
)
}
\description{
BUild the mysql database for the httr data
}
